+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/common/repository/GugunRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface GugunRepository extends JpaRepository<Gugun, Integer> {
+public interface GugunRepository extends JpaRepository<Gugun, Long> {
     @Query("SELECT g.id, g.name FROM Gugun g WHERE g.sidoId = :sidoId ORDER BY g.sidoId, g.id")
-    List<Object[]> findGugunByIdSidoOrderById(@Param("sidoId") Integer sidoId);
+    List<Object[]> findGugunByIdSidoOrderById(@Param("sidoId") Long sidoId);
 }
